+++ src/component/Schema/Jest/index.js
@@ -127,7 +127,7 @@ describe( ${ JSON.stringify( title ) }, () => {
     bs.page.on( "dialog", ( dialog ) => dialogLog.push( dialog ) );
 
     ${ options.requireInterceptTraffic ? `bs.performance.watchTraffic();` : `` }
-
+    
     ${ options.interactiveMode ? `
     let stepIndex = 0;
     await bs.page.exposeFunction('setPuppetryStepIndex', index => {