+++ src/component/Schema/schema.jsx
@@ -35,6 +35,7 @@ import { assertPerformanceAssetWeight } from "./Params/Page/assertPerformanceAss
 import { assertPerformanceAssetCount } from "./Params/Page/assertPerformanceAssetCount";
 import { assertPerformanceTiming } from "./Params/Page/assertPerformanceTiming";
 import { assertGaTracking } from "./Params/Page/assertGaTracking";
+import { assertResponse } from "./Params/Page/assertResponse";
 
 import { screenshot as screenshotElement } from "./Params/Element/screenshot";
 import { assertProperty } from "./Params/Element/assertProperty";
@@ -142,6 +143,7 @@ export const schema = {
     assertScroll: assertScrollPage,
     assertVar,
     assertScreenshot: assertScreenshotPage,
+    assertResponse,
     assertConsoleMessage,
     assertDialog,
     assertPerformanceAssetWeight,