+++ src/service/assert.js
@@ -175,6 +175,11 @@ function createCbBody({ assert, target, method, id, params }) {
     return justify( `expect( result )`
           + `.toMatchBoundingBoxSnapshot( ${ JSON.stringify( options, null, "  " ) }, "${ source }" );` );
 
+  case "response":
+    return justify( options.textOperator !== "any" ? `result.data = await result.text();\n` : `` )
+          + justify( `expect( result )`
+          + `.toMatchResponse( ${ JSON.stringify( params.url ) }, `
+          + `${ JSON.stringify( options, null, "  " ) }, "${ source }" );` );
   case "assertPerformanceTiming":
     return resolveTimingAssertion( options, source );
   case "assertAssetWeight":